+++ CHANGELOG.md
@@ -18,6 +18,10 @@ For a full diff see [`7afa59c...main`][7afa59c...main].
 * Added `Comparator\DurationComparator` ([#18]), by [@localheinz]
 * Added `SlowTestReporter` ([#19]), by [@localheinz]
 
+### Changed
+
+* Renamed `SlowTestReporter` to `Reporter\DefaultReporter` ([#20]), by [@localheinz]
+
 [7afa59c...main]: https://github.com/ergebnis/phpunit-slow-test-detector/compare/7afa59c...main
 
 [#6]: https://github.com/ergebnis/phpunit-slow-test-detector/pull/6
@@ -27,5 +31,6 @@ For a full diff see [`7afa59c...main`][7afa59c...main].
 [#17]: https://github.com/ergebnis/phpunit-slow-test-detector/pull/17
 [#18]: https://github.com/ergebnis/phpunit-slow-test-detector/pull/18
 [#19]: https://github.com/ergebnis/phpunit-slow-test-detector/pull/19
+[#20]: https://github.com/ergebnis/phpunit-slow-test-detector/pull/20
 
 [@localheinz]: https://github.com/localheinz